add `.env` file check in `initial-setup.sh`

+++ scripts/intial-setup.sh
@@ -1,18 +1,22 @@
 echo "This is 3 step process: Initializing the .env files, installing the dependencies and generating prisma types."
 
 # -----
-echo "creating .env file..."
 ENV_FILE="./apps/api/.env"
+if [ -f "$ENV_FILE" ]
+then
+    echo ".env file already exists at $ENV_FILE. Skipping..."
+else
+	echo "creating .env file..."
+    read -p "Enter the postgresSQL URL: " PGSQLURL
+    read -p "Enter a secret-key for JWT: " SECRETKEY
 
-read -p "Enter the postgresSQL URL: " PGSQLURL
-read -p "Enter a secret-key for JWT: " SECRETKEY
+    mv $ENV_FILE $ENV_FILE.bak
+    touch $ENV_FILE
 
-mv $ENV_FILE $ENV_FILE.bak
-touch $ENV_FILE
-
-echo "PGSQL_DB_URL=\"$PGSQLURL\"" > $ENV_FILE
-echo "JWT_SECRET=\"$SECRETKEY\"" >> $ENV_FILE
-echo "data appended to $ENV_FILE"
+    echo "PGSQL_DB_URL=\"$PGSQLURL\"" > $ENV_FILE
+    echo "JWT_SECRET=\"$SECRETKEY\"" >> $ENV_FILE
+    echo "data appended to $ENV_FILE"
+fi
 
 # -----
 echo "Installing npm packages..."
